+++ MonoGame.Framework/Graphics/SpriteBatch.cs
@@ -454,21 +454,21 @@ namespace Microsoft.Xna.Framework.Graphics
                 _texCoordBR.Y = (srcRect.Y + srcRect.Height) * texture.TexelHeight;
 
                 if(srcRect.Width != 0)
-                    origin.X = origin.X * (float)destinationRectangle.Width / (float)srcRect.Width;
+                    origin.X = origin.X * destinationRectangle.Width / srcRect.Width;
                 else
-                    origin.X = origin.X * (float)destinationRectangle.Width * texture.TexelWidth;
+                    origin.X = origin.X * destinationRectangle.Width * texture.TexelWidth;
                 if(srcRect.Height != 0)
-                    origin.Y = origin.Y * (float)destinationRectangle.Height / (float)srcRect.Height; 
+                    origin.Y = origin.Y * destinationRectangle.Height / srcRect.Height; 
                 else
-                    origin.Y = origin.Y * (float)destinationRectangle.Height * texture.TexelHeight;
+                    origin.Y = origin.Y * destinationRectangle.Height * texture.TexelHeight;
             }
             else
             {
                 _texCoordTL = Vector2.Zero;
                 _texCoordBR = Vector2.One;
                 
-                origin.X = origin.X * (float)destinationRectangle.Width  * texture.TexelWidth;
-                origin.Y = origin.Y * (float)destinationRectangle.Height * texture.TexelHeight;
+                origin.X = origin.X * destinationRectangle.Width * texture.TexelWidth;
+                origin.Y = origin.Y * destinationRectangle.Height * texture.TexelHeight;
             }
             
 			if ((effects & SpriteEffects.FlipVertically) != 0)
@@ -1145,10 +1145,10 @@ namespace Microsoft.Xna.Framework.Graphics
                     item.Texture = spriteFont.Texture;
                     item.SortKey = sortKey;
 
-                    _texCoordTL.X = pCurrentGlyph->BoundsInTexture.X * (float)spriteFont.Texture.TexelWidth;
-                    _texCoordTL.Y = pCurrentGlyph->BoundsInTexture.Y * (float)spriteFont.Texture.TexelHeight;
-                    _texCoordBR.X = (pCurrentGlyph->BoundsInTexture.X + pCurrentGlyph->BoundsInTexture.Width) * (float)spriteFont.Texture.TexelWidth;
-                    _texCoordBR.Y = (pCurrentGlyph->BoundsInTexture.Y + pCurrentGlyph->BoundsInTexture.Height) * (float)spriteFont.Texture.TexelHeight;
+                    _texCoordTL.X = pCurrentGlyph->BoundsInTexture.X * spriteFont.Texture.TexelWidth;
+                    _texCoordTL.Y = pCurrentGlyph->BoundsInTexture.Y * spriteFont.Texture.TexelHeight;
+                    _texCoordBR.X = (pCurrentGlyph->BoundsInTexture.X + pCurrentGlyph->BoundsInTexture.Width) * spriteFont.Texture.TexelWidth;
+                    _texCoordBR.Y = (pCurrentGlyph->BoundsInTexture.Y + pCurrentGlyph->BoundsInTexture.Height) * spriteFont.Texture.TexelHeight;
 
                     if ((effects & SpriteEffects.FlipVertically) != 0)
                     {
