+++ MonoGame.Framework/Media/MediaQueue.cs
@@ -70,7 +70,7 @@ namespace Microsoft.Xna.Framework.Media
 			if (shuffle)
 				_activeSongIndex = random.Next(songs.Count);
 			else			
-				_activeSongIndex = (int)MathHelper.Clamp(_activeSongIndex + direction, 0, songs.Count - 1);
+				_activeSongIndex = MathHelper.Clamp(_activeSongIndex + direction, 0, songs.Count - 1);
 			
 			return songs[_activeSongIndex];
 		}
